+++ examples/filter_advanced.py
@@ -3,7 +3,7 @@ import pandas as pd
 
 
 def name_filter(e):
-    return spp.filter.or_(e.FirstName == "Stefan", e.LastName == "Seiler")
+    return spp.Filter.or_(e.FirstName == "Stefan", e.LastName == "Seiler")
 
 
 persons = spp.get_data("Person", filter=name_filter, Language="DE")